+++ kie-dmn/kie-dmn-validation/src/test/java/org/kie/dmn/validation/dtanalysis/GapsAndOverlaps1Test.java
@@ -35,17 +35,37 @@ import static org.junit.Assert.assertThat;
 import static org.kie.dmn.validation.DMNValidator.Validation.ANALYZE_DECISION_TABLE;
 import static org.kie.dmn.validation.DMNValidator.Validation.VALIDATE_COMPILATION;
 
-public class GapsAndOverlaps1_domainOnTable_Test extends AbstractDTAnalysisTest {
+public class GapsAndOverlaps1Test extends AbstractDTAnalysisTest {
 
     @Test
     public void test() {
-        List<DMNMessage> validate = validator.validate(getReader("GapsAndOverlaps1-domainOnTable.dmn"), VALIDATE_COMPILATION, ANALYZE_DECISION_TABLE);
+        List<DMNMessage> validate = validator.validate(getReader("GapsAndOverlaps1.dmn"), VALIDATE_COMPILATION, ANALYZE_DECISION_TABLE);
         DTAnalysis analysis = getAnalysis(validate, "_cd2e0a28-3cc2-456b-90b6-392d9c3574af");
-
-        assertThat(analysis.getGaps(), hasSize(11));
+        
+        assertThat(analysis.getGaps(), hasSize(17));
 
         @SuppressWarnings({"unchecked", "rawtypes"})
         List<Hyperrectangle> gaps = Arrays.asList(new Hyperrectangle(2,
+                                                                     Arrays.asList(Interval.newFromBounds(new Bound(Interval.NEG_INF,
+                                                                                                                    RangeBoundary.CLOSED,
+                                                                                                                    null),
+                                                                                                          new Bound(new BigDecimal("0"),
+                                                                                                                    RangeBoundary.OPEN,
+                                                                                                                    null)))),
+                                                  new Hyperrectangle(2,
+                                                                     Arrays.asList(Interval.newFromBounds(new Bound(new BigDecimal("0"),
+                                                                                                                    RangeBoundary.CLOSED,
+                                                                                                                    null),
+                                                                                                          new Bound(new BigDecimal("1"),
+                                                                                                                    RangeBoundary.OPEN,
+                                                                                                                    null)),
+                                                                                   Interval.newFromBounds(new Bound(Interval.NEG_INF,
+                                                                                                                    RangeBoundary.CLOSED,
+                                                                                                                    null),
+                                                                                                          new Bound(new BigDecimal("0"),
+                                                                                                                    RangeBoundary.OPEN,
+                                                                                                                    null)))),
+                                                  new Hyperrectangle(2,
                                                                      Arrays.asList(Interval.newFromBounds(new Bound(new BigDecimal("0"),
                                                                                                                     RangeBoundary.CLOSED,
                                                                                                                     null),
@@ -55,7 +75,20 @@ public class GapsAndOverlaps1_domainOnTable_Test extends AbstractDTAnalysisTest 
                                                                                    Interval.newFromBounds(new Bound(new BigDecimal("2"),
                                                                                                                     RangeBoundary.OPEN,
                                                                                                                     null),
-                                                                                                          new Bound(new BigDecimal("999"),
+                                                                                                          new Bound(Interval.POS_INF,
+                                                                                                                    RangeBoundary.CLOSED,
+                                                                                                                    null)))),
+                                                  new Hyperrectangle(2,
+                                                                     Arrays.asList(Interval.newFromBounds(new Bound(new BigDecimal("1"),
+                                                                                                                    RangeBoundary.CLOSED,
+                                                                                                                    null),
+                                                                                                          new Bound(new BigDecimal("2"),
+                                                                                                                    RangeBoundary.OPEN,
+                                                                                                                    null)),
+                                                                                   Interval.newFromBounds(new Bound(Interval.NEG_INF,
+                                                                                                                    RangeBoundary.CLOSED,
+                                                                                                                    null),
+                                                                                                          new Bound(new BigDecimal("0"),
                                                                                                                     RangeBoundary.OPEN,
                                                                                                                     null)))),
                                                   new Hyperrectangle(2,
@@ -81,7 +114,20 @@ public class GapsAndOverlaps1_domainOnTable_Test extends AbstractDTAnalysisTest 
                                                                                    Interval.newFromBounds(new Bound(new BigDecimal("6"),
                                                                                                                     RangeBoundary.OPEN,
                                                                                                                     null),
-                                                                                                          new Bound(new BigDecimal("999"),
+                                                                                                          new Bound(Interval.POS_INF,
+                                                                                                                    RangeBoundary.CLOSED,
+                                                                                                                    null)))),
+                                                  new Hyperrectangle(2,
+                                                                     Arrays.asList(Interval.newFromBounds(new Bound(new BigDecimal("2"),
+                                                                                                                    RangeBoundary.CLOSED,
+                                                                                                                    null),
+                                                                                                          new Bound(new BigDecimal("3"),
+                                                                                                                    RangeBoundary.CLOSED,
+                                                                                                                    null)),
+                                                                                   Interval.newFromBounds(new Bound(Interval.NEG_INF,
+                                                                                                                    RangeBoundary.CLOSED,
+                                                                                                                    null),
+                                                                                                          new Bound(new BigDecimal("0"),
                                                                                                                     RangeBoundary.OPEN,
                                                                                                                     null)))),
                                                   new Hyperrectangle(2,
@@ -107,7 +153,20 @@ public class GapsAndOverlaps1_domainOnTable_Test extends AbstractDTAnalysisTest 
                                                                                    Interval.newFromBounds(new Bound(new BigDecimal("6"),
                                                                                                                     RangeBoundary.OPEN,
                                                                                                                     null),
-                                                                                                          new Bound(new BigDecimal("999"),
+                                                                                                          new Bound(Interval.POS_INF,
+                                                                                                                    RangeBoundary.CLOSED,
+                                                                                                                    null)))),
+                                                  new Hyperrectangle(2,
+                                                                     Arrays.asList(Interval.newFromBounds(new Bound(new BigDecimal("3"),
+                                                                                                                    RangeBoundary.CLOSED,
+                                                                                                                    null),
+                                                                                                          new Bound(new BigDecimal("4"),
+                                                                                                                    RangeBoundary.CLOSED,
+                                                                                                                    null)),
+                                                                                   Interval.newFromBounds(new Bound(Interval.NEG_INF,
+                                                                                                                    RangeBoundary.CLOSED,
+                                                                                                                    null),
+                                                                                                          new Bound(new BigDecimal("0"),
                                                                                                                     RangeBoundary.OPEN,
                                                                                                                     null)))),
                                                   new Hyperrectangle(2,
@@ -120,8 +179,8 @@ public class GapsAndOverlaps1_domainOnTable_Test extends AbstractDTAnalysisTest 
                                                                                    Interval.newFromBounds(new Bound(new BigDecimal("4"),
                                                                                                                     RangeBoundary.OPEN,
                                                                                                                     null),
-                                                                                                          new Bound(new BigDecimal("999"),
-                                                                                                                    RangeBoundary.OPEN,
+                                                                                                          new Bound(Interval.POS_INF,
+                                                                                                                    RangeBoundary.CLOSED,
                                                                                                                     null)))),
                                                   new Hyperrectangle(2,
                                                                      Arrays.asList(Interval.newFromBounds(new Bound(new BigDecimal("4"),
@@ -130,7 +189,7 @@ public class GapsAndOverlaps1_domainOnTable_Test extends AbstractDTAnalysisTest 
                                                                                                           new Bound(new BigDecimal("5"),
                                                                                                                     RangeBoundary.CLOSED,
                                                                                                                     null)),
-                                                                                   Interval.newFromBounds(new Bound(new BigDecimal("0"),
+                                                                                   Interval.newFromBounds(new Bound(Interval.NEG_INF,
                                                                                                                     RangeBoundary.CLOSED,
                                                                                                                     null),
                                                                                                           new Bound(new BigDecimal("1"),
@@ -146,8 +205,8 @@ public class GapsAndOverlaps1_domainOnTable_Test extends AbstractDTAnalysisTest 
                                                                                    Interval.newFromBounds(new Bound(new BigDecimal("4"),
                                                                                                                     RangeBoundary.OPEN,
                                                                                                                     null),
-                                                                                                          new Bound(new BigDecimal("999"),
-                                                                                                                    RangeBoundary.OPEN,
+                                                                                                          new Bound(Interval.POS_INF,
+                                                                                                                    RangeBoundary.CLOSED,
                                                                                                                     null)))),
                                                   new Hyperrectangle(2,
                                                                      Arrays.asList(Interval.newFromBounds(new Bound(new BigDecimal("5"),
@@ -163,11 +222,24 @@ public class GapsAndOverlaps1_domainOnTable_Test extends AbstractDTAnalysisTest 
                                                                                                           new Bound(new BigDecimal("7"),
                                                                                                                     RangeBoundary.CLOSED,
                                                                                                                     null)),
+                                                                                   Interval.newFromBounds(new Bound(Interval.NEG_INF,
+                                                                                                                    RangeBoundary.CLOSED,
+                                                                                                                    null),
+                                                                                                          new Bound(new BigDecimal("0"),
+                                                                                                                    RangeBoundary.OPEN,
+                                                                                                                    null)))),
+                                                  new Hyperrectangle(2,
+                                                                     Arrays.asList(Interval.newFromBounds(new Bound(new BigDecimal("6"),
+                                                                                                                    RangeBoundary.CLOSED,
+                                                                                                                    null),
+                                                                                                          new Bound(new BigDecimal("7"),
+                                                                                                                    RangeBoundary.CLOSED,
+                                                                                                                    null)),
                                                                                    Interval.newFromBounds(new Bound(new BigDecimal("3"),
                                                                                                                     RangeBoundary.OPEN,
                                                                                                                     null),
-                                                                                                          new Bound(new BigDecimal("999"),
-                                                                                                                    RangeBoundary.OPEN,
+                                                                                                          new Bound(Interval.POS_INF,
+                                                                                                                    RangeBoundary.CLOSED,
                                                                                                                     null)))),
                                                   new Hyperrectangle(2,
                                                                      Arrays.asList(Interval.newFromBounds(new Bound(new BigDecimal("7"),
@@ -176,9 +248,9 @@ public class GapsAndOverlaps1_domainOnTable_Test extends AbstractDTAnalysisTest 
                                                                                                           new Bound(Interval.POS_INF,
                                                                                                                     RangeBoundary.CLOSED,
                                                                                                                     null)))));
-        assertThat(gaps, hasSize(11));
+        assertThat(gaps, hasSize(17));
 
-        // Assert GAPs same values
+        // Assert GAPS
         assertThat(analysis.getGaps(), contains(gaps.toArray()));
 
         // assert OVERLAPs count.
